requirements.txt
asgiref==3.5.0
asttokens==2.0.5
atomicwrites==1.4.1        
attrs==21.4.0
backcall==0.2.0
boto3==1.24.7
botocore==1.27.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12 
click==8.1.2
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1       
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.3        
Django==4.0.2
django-allauth==0.47.0     
django-cors-headers==3.11.0
django-countries==7.3.2    
django-crontab==0.7.1
django-dbbackup==4.0.0b0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-filter==21.1
django-jsonform==2.10.1
django-storages==1.12.3
django-tenants==3.4.2
djangorestframework==3.13.1
djangorestframework-gis==0.18
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
drf-yasg==1.20.0
executing==0.8.3
factory-boy==3.2.1
Faker==13.15.1
freezegun==1.2.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
ipython==8.1.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.0
jsonschema==4.4.0
kml2geojson==5.1.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
pygeoif==0.7
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
Shapely==1.8.1.post1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
swagger-spec-validator==2.7.4
tomli==2.0.1
tomlkit==0.11.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
zipp==3.8.1
